+++ include/special/Page.php
@@ -109,7 +109,7 @@ class Page extends \gp\Page{
 		}
 
 		if( \gp\admin\Tools::HasPermission('Admin_User') ){
-			$option_links[] = \gp\tool::Link('Admin/Users',$langmessage['permissions'],'cmd=file_permissions&index='.urlencode($this->gp_index),array('title'=>$langmessage['permissions'],'data-cmd'=>'gpabox'));
+			$option_links[] = \gp\tool::Link('Admin/Permissions',$langmessage['permissions'],'index='.urlencode($this->gp_index),array('title'=>$langmessage['permissions'],'data-cmd'=>'gpabox'));
 		}
 
 		if( !empty($option_links) ){
